pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "requests>=2.32.4",
    "whitenoise>=6.9.0",
]
